requirements.txt
anyio==4.11.0
asgiref==3.9.2
asyncio==4.0.0
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudscraper==1.2.71
colorama==0.4.6
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
pycparser==2.23
pyparsing==3.2.5
PySocks==1.7.1
python-telegram-bot==22.5
requests==2.32.5
requests-toolbelt==1.0.0
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
gunicorn